Look for ninja-build and ninja.
Fedora and CentOS have taken to calling the ninja executable
"ninja-build", meaning cninja doesn't work out of the box.  It's easy
enough to create a symlink somewhere to fix the problem, but it seemed
like a better idea to make cninja work out-of-the-box.

This does introduce the use of the File::Which package to help search
the path for the binary, which adds an extra dependency.